Outsourcing of connection stuff in own file
